requirements.txt
Django==2.2.6
django-crontab==0.7.1
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
